refactor: inline ScheduleManagementContent, drop .impl.tsx and scroll comments
