+++ apps/antalmanac/src/components/ScheduleManagement/ScheduleManagement.tsx
@@ -26,15 +26,10 @@ export function ScheduleManagement() {
         }))
     );
 
-    // Tab name mapped to the last known scrollTop.
     const [positions, setPositions] = useState<Partial<Record<TabName, number>>>({});
 
-    /**
-     * Ref to the scrollable container with all of the tabs-content within it.
-     */
     const ref = useRef<HTMLDivElement>(null);
 
-    // Save the current scroll position to the store.
     const onScroll = (e: React.UIEvent<HTMLDivElement, UIEvent>) => {
         const positionToSave = e.currentTarget.scrollTop;
         setPositions((current) => {
@@ -68,7 +63,6 @@ export function ScheduleManagement() {
         }
     }, [segment, fallbackMode, router]);
 
-    // Restore scroll position if it has been previously saved.
     useEffect(() => {
         const savedPosition = positions[activeTab];
